CAMS-770 Move division filter helpers to court-utils
Moves autoUpgradeToAll, resolveCombinedSelections, getUserDivisionCodes,
and encodeDivisionCodes from trusteeDistrictFilterUseCase and
trusteeCaseListFilterUseCase into court-utils so that the shared
DistrictDivisionComboBox component can import them without crossing a
feature module boundary. Removes the debug console.log from
TrusteeDistrictFilterView.

Jira ticket: CAMS-770
